requirements.txt
certifi==2019.6.16
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.14
numpy==1.16.4
pandas==0.24.2
protobuf==3.8.0
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
Werkzeug==0.15.4
wincertstore==0.2
